# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sales System API",
    "description": "نظام المبيعات المتقدم — API شامل لإدارة الحسابات والمعاملات المالية والموظفين والرواتب",
    "version": "1.0.0",
    "contact": { "name": "Sales System Team" }
  },
  "servers": [
    { "url": "/api", "description": "API Base" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token من POST /v1/auth/login"
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/v1/backups": {
      "get": {
        "tags": ["Backups"],
        "summary": "عرض النسخ الاحتياطية (آخر 20)",
        "responses": {
          "200": { "description": "قائمة النسخ الاحتياطية" }
        }
      },
      "post": {
        "tags": ["Backups"],
        "summary": "إنشاء نسخة احتياطية جديدة",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["data"],
                "properties": {
                  "data": { "type": "object", "description": "البيانات المطلوب نسخها احتياطياً" },
                  "label": { "type": "string", "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "تم إنشاء النسخة الاحتياطية" }
        }
      }
    },
    "/v1/backups/{backupId}": {
      "get": {
        "tags": ["Backups"],
        "summary": "عرض تفاصيل نسخة احتياطية",
        "parameters": [
          { "name": "backupId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "تفاصيل النسخة الاحتياطية" },
          "404": { "description": "النسخة غير موجودة" }
        }
      },
      "delete": {
        "tags": ["Backups"],
        "summary": "حذف نسخة احتياطية",
        "parameters": [
          { "name": "backupId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "تم الحذف بنجاح" },
          "404": { "description": "النسخة غير موجودة" }
        }
      }
    },
...
  }
}